[chore] Changed import statements in run module and moved all project file to src
